Backend/python_llm/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
motor
beautifulsoup4
langchain
openai
anyio
